refactor: model move to domain
